Init dagster refactor

+++ src/pudl/load.py
@@ -7,6 +7,7 @@ from sqlite3 import sqlite_version
 
 import pandas as pd
 import sqlalchemy as sa
+from dagster import Field, op
 from packaging import version
 from sqlalchemy.exc import IntegrityError
 
@@ -18,16 +19,23 @@ logger = logging.getLogger(__name__)
 MINIMUM_SQLITE_VERSION = "3.32.0"
 
 
+@op(
+    config_schema={
+        "check_foreign_keys": Field(bool, default_value=True),
+        "check_types": Field(bool, default_value=True),
+        "check_values": Field(bool, default_value=True),
+        "clobber": Field(bool, default_value=True),
+    },
+    required_resource_keys={"pudl_engine"},
+)
 def dfs_to_sqlite(
+    context,
     dfs: dict[str, pd.DataFrame],
-    engine: sa.engine.Engine,
-    check_foreign_keys: bool = True,
-    check_types: bool = True,
-    check_values: bool = True,
 ) -> None:
     """Load a dictionary of dataframes into the PUDL SQLite DB.
 
     Args:
+        context: dagster context keyword.
         dfs: Dictionary mapping table names to dataframes.
         engine: PUDL DB connection engine.
         check_foreign_keys: if True, enforce foreign key constraints.
@@ -35,8 +43,11 @@ def dfs_to_sqlite(
         check_values: if True, enforce value constraints.
 
     """
+    engine = context.resources.pudl_engine
+    check_foreign_keys = context.op_config["check_foreign_keys"]
     # This magic makes SQLAlchemy tell SQLite to check foreign key constraints
     # whenever we insert data into thd database, which it doesn't do by default
+
     @sa.event.listens_for(sa.engine.Engine, "connect")
     def _set_sqlite_pragma(dbapi_connection, connection_record):
         if isinstance(dbapi_connection, SQLite3Connection):
@@ -49,6 +60,7 @@ def dfs_to_sqlite(
     bad_sqlite_version = version.parse(sqlite_version) < version.parse(
         MINIMUM_SQLITE_VERSION
     )
+    check_types = context.op_config["check_types"]
     if bad_sqlite_version and check_types:
         check_types = False
         logger.warning(
@@ -59,10 +71,12 @@ def dfs_to_sqlite(
 
     # Generate a SQLAlchemy MetaData object from dataframe names:
     md = Package.from_resource_ids(resource_ids=tuple(sorted(dfs))).to_sql(
-        check_types=check_types, check_values=check_values
+        check_types=check_types,
+        check_values=context.op_config["check_values"],
     )
     # Delete any existing tables, and create them anew:
-    md.drop_all(engine)
+    if context.op_config["clobber"]:
+        md.drop_all(engine)
     md.create_all(engine)
 
     # Load any tables that exist in our dictionary of dataframes into the
